feat: [CDS-101927]: moved Pipeline Resource to new folder
