+++ internal/provider/provider.go
@@ -7,7 +7,7 @@ import (
 	"github.com/harness/terraform-provider-harness/internal/service/platform/service_account"
 	"log"
 
-	pipeline_gitx "github.com/harness/terraform-provider-harness/internal/service/cd_nextgen/gitx/webhook"
+	pipeline_gitx "github.com/harness/terraform-provider-harness/internal/service/pipeline/gitx/webhook"
 	"github.com/harness/terraform-provider-harness/internal/service/platform/cluster_orchestrator"
 	dbinstance "github.com/harness/terraform-provider-harness/internal/service/platform/db_instance"
 	dbschema "github.com/harness/terraform-provider-harness/internal/service/platform/db_schema"
@@ -81,15 +81,15 @@ import (
 	gitops_repo_cert "github.com/harness/terraform-provider-harness/internal/service/platform/gitops/repository_certificates"
 	gitops_repo_cred "github.com/harness/terraform-provider-harness/internal/service/platform/gitops/repository_credentials"
 
-	pipeline_input_set "github.com/harness/terraform-provider-harness/internal/service/cd_nextgen/input_set"
 	cdng_overrides "github.com/harness/terraform-provider-harness/internal/service/cd_nextgen/overrides"
-	pipeline "github.com/harness/terraform-provider-harness/internal/service/cd_nextgen/pipeline"
-	pipeline_filters "github.com/harness/terraform-provider-harness/internal/service/cd_nextgen/pipeline_filters"
 	cdng_service_overrides_v2 "github.com/harness/terraform-provider-harness/internal/service/cd_nextgen/service_overrides_v2"
-	pipeline_template "github.com/harness/terraform-provider-harness/internal/service/cd_nextgen/template"
-	pipeline_template_filters "github.com/harness/terraform-provider-harness/internal/service/cd_nextgen/template_filters"
-	pipeline_triggers "github.com/harness/terraform-provider-harness/internal/service/cd_nextgen/triggers"
 	cdng_variables "github.com/harness/terraform-provider-harness/internal/service/cd_nextgen/variables"
+	pipeline_input_set "github.com/harness/terraform-provider-harness/internal/service/pipeline/input_set"
+	"github.com/harness/terraform-provider-harness/internal/service/pipeline/pipeline"
+	"github.com/harness/terraform-provider-harness/internal/service/pipeline/pipeline_filters"
+	pipeline_template "github.com/harness/terraform-provider-harness/internal/service/pipeline/template"
+	pipeline_template_filters "github.com/harness/terraform-provider-harness/internal/service/pipeline/template_filters"
+	pipeline_triggers "github.com/harness/terraform-provider-harness/internal/service/pipeline/triggers"
 	"github.com/harness/terraform-provider-harness/internal/service/platform/monitored_service"
 	"github.com/harness/terraform-provider-harness/internal/service/platform/organization"
 	pl_permissions "github.com/harness/terraform-provider-harness/internal/service/platform/permissions"
