+++ internal/service/pipeline/template/resource_template.go
